# src/docs/swaggerConfig.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger ECommerce",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "security": {
    "cookieAuth": []
  },
  "paths": {
    "/api/products": {
      "get": {
        "summary": "List all Products",
        "operationId": "listProducts",
        "tags": [
          "products"
        ],
        "responses": {
          "200": {
            "description": "Every product available",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Products"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a Product",
        "operationId": "createProduct",
        "tags": [
          "products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create Product",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Null response"
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
    },
...
    "/api/cart/{cartId}": {
      "get": {
        "summary": "Info for a specific Cart",
        "operationId": "showCartById",
        "tags": [
          "carts"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "description": "The id of the cart to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cart by Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific Cart",
        "operationId": "createCart",
        "tags": [
          "carts"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "required": true,
            "description": "The id of the cart to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Cart",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Cart"
                  }
                }
              }
            }
          },
          "201": {
            "description": "Null response"
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cookie.id"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
            "id",
            "timestamp",
            "name",
            "description",
            "code",
            "url",
            "price",
            "stock"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "ObjectId"
          },
          "timestamp": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          }
        }
      },
      "ProductDTO": {
        "type": "object",
        "required": [
            "name",
            "description",
            "code",
            "url",
            "price",
            "stock"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          }
        },
        "example": {
          "name": "TV",
          "description": "Television LED",
          "code": "TV159",
          "url": "www.compumundo.com",
          "price": 80000,
          "stock": 10
        }
      },
      "Products": {
        "type": "array",
        "maxItems": 100,
        "items": {
          "$ref": "#/components/schemas/Product"
        }
      },
      "Cart": {
        "type": "object",
        "required": [
            "id",
            "timestamp",
            "products"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "ObjectId"
          },
          "timestamp": {
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}